The testdata is organize using d3.nest(), and then filter by date
Using d3.nest(), we set the key by a post's created_time following the
format: 'year-month'. In the meantime, we can filter a user's posts
based on the 'year-month' format. For example, if we want to visualize
all the posts in March 2016, we can simply define the display_date to
'2016-3'.

todos:

  Fix the clipping problem of the date string on timeline.

  Fix the clipping problem of the text box next to the timeline.

  Query different time's post on visualization or webpage, using
  pagination, Infinite Scroll, date picker and so on.